+++ lib/patch_command.dart
@@ -1,10 +1,12 @@
 import 'dart:convert';
 import 'dart:io';
 
+import 'package:args/command_runner.dart';
+import 'package:dcli/dcli.dart';
 import 'package:path/path.dart' as p;
 
-import 'package:args/args.dart';
-import 'package:args/command_runner.dart';
+import 'dart_import_app.dart';
+import 'src/version/version.g.dart';
 
 class PatchCommand extends Command<void> {
   @override
@@ -14,12 +16,57 @@ class PatchCommand extends Command<void> {
   @override
   String get name => 'patch';
 
+  DartProject _project;
+
+  String _projectRoot;
+
+  PatchCommand() {
+    argParser.addOption('root',
+        help: 'The path to the root of your project.', valueHelp: 'path');
+    argParser.addFlag('debug',
+        defaultsTo: false, negatable: false, help: 'Turns on debug ouput');
+    argParser.addFlag('version',
+        abbr: 'v',
+        defaultsTo: false,
+        negatable: false,
+        help: 'Outputs the version of drtimport and exits.');
+  }
+
   @override
   void run() {
-    if (argResults.rest.length != 2) {
-      fullusage(argParser);
-      exit(-1);
+    var root = argResults['root'] as String;
+
+    root ??= pwd;
+
+    if (argResults['version'] == true) {
+      fullusage();
     }
+
+    if (argResults['debug'] == true) DartImportApp().enableDebug();
+
+    if (argResults.rest.length != 2) {
+      fullusage();
+    }
+
+    _project = DartProject.fromPath(root);
+    _projectRoot = _project.pathToProjectRoot;
+
+    if (!_project.hasPubSpec) {
+      fullusage(
+          error:
+              'The project root directory ${_projectRoot} does not contain a pubspec.yaml');
+    }
+
+    var pubspec = _project.pubSpec;
+    print(orange('Processing project ${pubspec.name} in ${_projectRoot}'));
+
+    // check we are in the root.
+    if (!exists(join(_projectRoot, 'lib'))) {
+      fullusage(
+          error:
+              'Your project structure looks invalid. You must have a "lib" directory in the root of your project.');
+    }
+
     final fromPattern = argResults.rest[0];
     final toPattern = argResults.rest[1];
 
@@ -27,27 +74,25 @@ class PatchCommand extends Command<void> {
   }
 
   void process(String fromPattern, String toPattern) async {
-    final cwd = Directory('.');
-
-    final files = cwd.list(recursive: true);
-
     final dartFiles =
-        await files.where((file) => file.path.endsWith('.dart')).toList();
+        find('*.dart', root: _project.pathToProjectRoot, recursive: true)
+            .toList();
 
     var scanned = 0;
     var updated = 0;
     for (var file in dartFiles) {
       scanned++;
-      final result = await replaceString(file, fromPattern, toPattern);
-      final tmpFile = result.tmpFile;
+      final result = await replaceString(File(file), fromPattern, toPattern);
+      final tmpFile = result.pathToTempFile;
 
       if (result.changeCount != 0) {
         updated++;
-        final backupFile = await file.rename(file.path + '.bak');
-        await tmpFile.rename(file.path);
-        await backupFile.delete();
+        var backupFile = '$file.bak';
+        move(file, backupFile);
+        move(tmpFile, file);
+        delete(backupFile);
 
-        print('Updated : ${file.path} changed ${result.changeCount} lines');
+        print('Updated : ${file} changed ${result.changeCount} lines');
       }
     }
     print('Finished: scanned $scanned updated $updated');
@@ -58,12 +103,12 @@ class PatchCommand extends Command<void> {
     final systemTempDir = Directory.systemTemp;
 
     final tmpPath = p.join(systemTempDir.path, file.path);
-    final tmpFile = File(tmpPath);
+    final tmpFile = tmpPath;
 
-    final tmpDir = Directory(tmpFile.parent.path);
-    await tmpDir.create(recursive: true);
+    final tmpDir = p.dirname(tmpFile);
+    await createDir(tmpDir, recursive: true);
 
-    final tmpSink = tmpFile.openWrite();
+    final tmpSink = File(tmpFile).openWrite();
 
     final result = Result(tmpFile);
 
@@ -94,18 +139,21 @@ class PatchCommand extends Command<void> {
     return changeCount;
   }
 
-  void fullusage(ArgParser parser) {
+  void fullusage({String error}) {
+    print('drtimport version: ${packageVersion}');
     print('Usage: ');
     print(description);
     print('<from string> <to string>');
     print('e.g. AppClass app_class');
-    print(parser.usage);
+    print(argParser.usage);
+
+    exit(-1);
   }
 }
 
 class Result {
-  File tmpFile;
+  String pathToTempFile;
   int changeCount = 0;
 
-  Result(this.tmpFile);
+  Result(this.pathToTempFile);
 }